# add config url trailing slash
following the opening of this issue (https://github.com/DivanteLtd/vue-storefront/issues/4684) I created a configuration ( urlTrailingSlash ) to add or not the slash at the end of a url.

I changed the normalizeUrlPath function so that depending on the configuration the slash is added or not.

I also changed the prepareDynamicRoute function and added the pathToRegexpOptions attribute with the strict option following the slash config. pathToRegexpOptions tells the vuejs router not to remove the slash.