Fixed: Order history pagination (#4599)
the getOrdersHistory, which previously used to fetch all the orders, after introducing pagination support in "getOrdersHistory", it only fetches 20 orders, which is its default value.

Still, the order history page calls it only once, trying to fetch all the results, and on those results, it applies lazyLoadOrdersOnScroll.

Due to which the user can not see more than 20 orders.