Fix: Duplicate records on order histroy and Updating of order history on placing new orders(#4599)
Created a new action for user state to append orders in the state, which will only get calls from pagination and remove the previously added appending logic from refreshOrderHistory.
We had added appending logic for pagination on getOrderHistory, it fetches order history and appends to the state, if state already had any orders, but it created these issues:

On every window refresh, getOrderHistory runs and appends the same orders again into the state.

On placing the order,  getOrderHistory runs and appends the new order again into the state.

So, For pagination we wrote another action only append orders and will only run from pagination and we removed the changes from getOrderHistory ( we added to append the state, in case of older orders in state) so that older action will always refresh the state as we wrote a new action which will only append the state and only get call from pagination.