ENG-171 Run GitGuardian in CI as part of Unified CI in Integrations
If you add the GitGuardian app globally to the org, it only scans code
when a PR is opened. This doesn't protect us against the case where
someone just creates a branch with plaintext credentials, *but doesn't
create a PR*.

Adding GitGuardian scans as a GitHub action with a "on push" trigger
will make GitGuardian scan all code that appears in the repo, rather
than just code which is part of a PR.

To make this work, you will need to add GITGUARDIAN_API_KEY GitHub
secret in each repository. Organization-wide GitHub secrets apply only on private
repositories, so to cover all repos reliably you'll need to create a regular
per-repo secret, not an org-wide one.