Don't cache TS results
This is EXTREMELY annyoing that if you run build once and it fails,
the failed build will be cached. Now even if you fix the underlying
issue and rebuild, the result will still be cached and you'll be
getting a build error even though you fixed everything.

This wastes a lot of time and it must be the third time I've come
across this.